+++ frontend/src/components/EditWorkoutNameButton.tsx
@@ -6,6 +6,7 @@ import { EditTwoTone } from "@ant-design/icons";
 import { getWorkouts, updateWorkoutName } from "../api/api";
 import { AuthContext } from "../contexts/AuthProvider";
 import { useHistory } from "react-router-dom";
+import { changeNameToUrl, buttonClickTrue } from "../utils/utils";
 
 export interface IWorkout {
 	name: string;
@@ -16,81 +17,43 @@ export interface IWorkouts {
 	workouts: IWorkout[];
 }
 
-let workouts2 = workouts;
-
-const toCamelCase = (phrase: string): string => {
-	let newStr = "";
-	let ch1 = phrase[0];
-	if ((ch1 >= "a" && ch1 <= "z") || (ch1 >= "A" && ch1 <= "Z")) {
-		newStr = ch1.toLowerCase();
-	}
-
-	const capitalizeChAfterSpace = (ch: string) => {
-		if ((ch >= "a" && ch <= "z") || (ch >= "A" && ch <= "Z")) {
-			if (ch === ch.toUpperCase()) {
-				newStr += ch;
-			} else {
-				newStr += ch.toUpperCase();
-			}
-		} else if (ch >= "0" && ch <= "9") {
-			newStr += ch;
-		}
-	};
-
-	for (let i = 1; i < phrase.length; i++) {
-		const ch = phrase[i];
-		if (phrase[i - 1] == " ") {
-			const capitalizeCh = capitalizeChAfterSpace(ch);
-		} else if ((ch >= "a" && ch <= "z") || (ch >= "0" && ch <= "9")) {
-			newStr += ch;
-		} else if (ch >= "A" && ch <= "Z") {
-			newStr += ch.toLowerCase();
-		}
-	}
-	return newStr;
-};
-
 export const EditWorkoutNameButton: React.FC<{
 	oldWorkout: IWorkoutNameUrl;
 }> = ({ oldWorkout }) => {
-	const [workoutName, setWorkoutName] = useState<string>(oldWorkout.name);
+	const [oldWorkoutName, setOldWorkoutName] = useState<string>(
+		oldWorkout.name
+	);
 	const [newWorkoutName, setNewWorkoutName] = useState<string>("");
-	const [submittedWorkoutName, setSubmittedWorkoutName] =
-		useState<string>("");
-	const [oldWorkoutUrl, setOldWorkoutUrl] = useState<string>(oldWorkout.url);
+	// const [submittedWorkoutName, setSubmittedWorkoutName] =
+	// 	useState<string>("");
 	const [newWorkoutUrl, setNewWorkoutUrl] = useState<string>("");
-	const [editName, setEditName] = useState<boolean>(false);
-	const { userId } = useContext(AuthContext);
-
-	const changeNameToUrl = (workoutName: string) => {
-		let newUrl = toCamelCase(workoutName);
-		setNewWorkoutUrl(newUrl);
-	};
-
-	useEffect(() => {
-		console.log("update OG name");
-		setWorkoutName(oldWorkout.name);
-	}, [workoutName]);
-
+	const [editName, setEditName] = useState<boolean>(false); // is edit button clicked
 	const history = useHistory();
+
+
+	// whenever oldWorkout, one being diosplayed, pre-edit, is entered, 
+	useEffect(() => {
+		setOldWorkoutName(oldWorkout.name);
+	}, [oldWorkoutName]);
 
 	const handleSubmit = async (e: any) => {
 		e.preventDefault();
 		try {
 			if (newWorkoutUrl.length > 0) {
 				const workoutNameUpdated = await updateWorkoutName(
-					oldWorkoutUrl,
+					oldWorkout.url,
 					newWorkoutUrl,
 					newWorkoutName
 				);
 				if (workoutNameUpdated) {
+					// returned updated workout
 					console.log("updated workout data:", workoutNameUpdated);
 					history.push(`/workouts/${workoutNameUpdated.url}`);
 				}
 				setNewWorkoutUrl("");
 				setNewWorkoutName("");
+				setOldWorkoutName(newWorkoutName); 
 				setEditName(false);
-				setWorkoutName(newWorkoutName);
 			}
 		} catch (error) {
 			console.log("error updated workout name", error);
@@ -98,18 +61,14 @@ export const EditWorkoutNameButton: React.FC<{
 		}
 	};
 
-	const revealEditName = () => {
-		setEditName(true);
-	};
-
 	return (
 		<>
 			{!editName ? (
 				<h1>
-					{oldWorkout.name}
+					{oldWorkoutName}
 					<EditTwoTone
 						className="new-workout-edit-icon"
-						onClick={revealEditName}
+						onClick={() => setEditName(buttonClickTrue)}
 					/>
 				</h1>
 			) : (
@@ -117,10 +76,12 @@ export const EditWorkoutNameButton: React.FC<{
 					<Input
 						onChange={(e) => {
 							setNewWorkoutName(e.target.value),
-								changeNameToUrl(e.target.value);
+								setNewWorkoutUrl(
+									changeNameToUrl(e.target.value)
+								);
 						}}
 						value={newWorkoutName}
-						placeholder={workoutName}
+						placeholder={oldWorkoutName}
 						className="new-workout-input"
 					/>
 					<Button type="primary" onClick={handleSubmit}>
